+++ src/routes/Header.svelte
@@ -10,6 +10,8 @@
   import { showBanner, timelineFilter } from "$lib/stores/globalRunes.svelte";
   import { untrack } from "svelte";
   import { mainMenuItems } from "./menu";
+  import HomeOptions from "./HomeOptions.svelte";
+  import GlobalOptions from "./GlobalOptions.svelte";
 
   let _showBanner: boolean = $state(showBanner.get());
 
@@ -28,32 +30,6 @@
     });
 
     return matchedItem || null;
-  });
-
-  const optionsArr = [
-    ["0", $_("filter.canversation.all")],
-    ["1", $_("filter.canversation.onlyFollowee")],
-    ["2", $_("filter.canversation.none")],
-  ];
-  const selected = writable<string>(
-    optionsArr[timelineFilter.get().selectCanversation][0]
-  );
-
-  const {
-    elements: { root, item, hiddenInput },
-    helpers: { isChecked },
-  } = createRadioGroup({
-    defaultValue: optionsArr[timelineFilter.get().selectCanversation][0],
-    value: selected,
-  });
-  // $: console.log(timelineFilter.get.adaptMute);
-  selected.subscribe((value) => {
-    if (value !== undefined && value !== null) {
-      timelineFilter.update((cur) => {
-        console.log(cur);
-        return { ...cur, selectCanversation: Number(value) };
-      });
-    }
   });
 
   $effect(() => {
@@ -101,38 +77,9 @@
             {#snippet popoverContent()}
               <ul class="w-[320px] max-w-full flex flex-col">
                 {#if currentPage?.alt === "home"}
-                  <li class="mb-2">
-                    <div class="label">
-                      {$_("filter.menu.canversation")}
-                    </div>
-                    <div
-                      use:melt={$root}
-                      class="text-sm my-1 gap-1 flex flex-col data-[orientation=horizontal]:flex-row"
-                      aria-label="View density"
-                    >
-                      {#each optionsArr as [index, option]}
-                        <div class="flex items-center gap-3">
-                          <button
-                            use:melt={$item(index)}
-                            class="grid h-6 w-6 cursor-default place-items-center rounded-full border border-magnum-400 shadow-sm
-        hover:bg-magnum-800"
-                            id={option}
-                            aria-labelledby="{option}-label"
-                          >
-                            {#if $isChecked(index)}
-                              <div
-                                class="h-3 w-3 rounded-full bg-magnum-400"
-                              ></div>
-                            {/if}
-                          </button>
-                          <label for={option} id="{option}-label">
-                            {option}
-                          </label>
-                        </div>
-                      {/each}
-                      <input name="line-height" use:melt={$hiddenInput} />
-                    </div>
-                  </li>
+                  <HomeOptions />
+                {:else if currentPage?.alt === "global"}
+                  <GlobalOptions />
                 {/if}
                 <li>
                   <label class="label">
