+++ src/routes/global/GlobalTimeline.svelte
@@ -9,6 +9,8 @@
   import { createRxForwardReq } from "rx-nostr";
   import { now, type EventPacket } from "rx-nostr/src";
   import { onDestroy, onMount } from "svelte";
+  import * as Nostr from "nostr-typedef";
+  import { followList, timelineFilter } from "$lib/stores/globalRunes.svelte";
 
   let isOnMount = false;
   let amount = 50;
@@ -72,6 +74,14 @@
     }
     resetUniq?.();
   });
+
+  const checkGlobalFolloweePost = (note: Nostr.Event): boolean => {
+    if (timelineFilter.get().globalExcludeFollowee) {
+      return !followList.get().has(note.pubkey);
+    } else {
+      return true;
+    }
+  };
 </script>
 
 {#if since && globalRelays.length > 0}
@@ -97,6 +107,9 @@
     {amount}
     {tieKey}
     relays={globalRelays}
+    eventFilter={(note) => {
+      return checkGlobalFolloweePost(note);
+    }}
   >
     {#snippet content({ events, len })}
       <!-- <SetRepoReactions /> -->
